Resolve issue #31
Supports independent subcommands. This is especially useful for commands
like /world where there could be a lot of subcommands and options. Each
command has a vector of commands that subcommands can be added to. If a
command wants to evaluate its subcommands it has to call
HandleSubcommands(). This is done automatically for Lua commands (see
LuaCommand.hpp).